% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_def1}
\alias{class_def1}
\title{Weather classifications, definition 1}
\format{
A data frame with 3 columns:
\describe{
  \item{prognosticator_city}{Where the prognosticator is located}
  \item{year}{Classification year}
  \item{class}{Classification: early spring or long winter}
}
}
\source{
<https://open-meteo.com/en/docs/historical-weather-api>
}
\usage{
class_def1
}
\description{
A table listing early spring/long winter classifications for each prognosticator's city by year. The definition of "early spring" follows previous analyses by NOAA and 538 in classifying a year as "early spring" if the average temperature in February OR March is above the 15-year running average for the location.
}
\keyword{datasets}
