% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prognosticators}
\alias{prognosticators}
\title{Prognosticators data}
\format{
A data frame with over 300 rows and 14 columns:
\describe{
  \item{prognosticator_name}{Prognosticator name}
  \item{Status}{Active/Inactive status}
  \item{prognosticator_slug}{Unique identifier derived from the prognosticator's URL on Countdown to Groundhog Day. Use this for joining with predictions.}
  \item{prognosticator_type_orig}{Prognosticator type, original labels from source}
  \item{prognosticator_city}{Where the prognosticator is located}
  \item{Last Prediction}{Year of last recorded prediction}
  \item{prognosticator_lat}{City latitude}
  \item{prognosticator_long}{City longitude}
  \item{prognosticator_type}{Prognosticator type, cleaned}
  \item{prognosticator_status}{Prognosticator status: creature, human mascot, inanimate}
  \item{prognosticator_creature}{Prognosticator creature type}
  \item{prognosticator_phylum}{Prognosticator creature phylum}
  \item{prognosticator_class}{Prognosticator creature class}
  \item{prognosticator_order}{Prognosticator creature order}
}
}
\source{
<https://countdowntogroundhogday.com/groundhogs-from-around-the-world>
}
\usage{
prognosticators
}
\description{
A table listing all known February 2nd prognosticators
}
\keyword{datasets}
