% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather_stations_ghcnd}
\alias{weather_stations_ghcnd}
\title{NOAA GHCND Weather Stations Near Prognosticators}
\format{
A data frame with 15045 rows and 6 columns:
\describe{
  \item{prognosticator_city}{Where the prognosticator is located}
  \item{id}{The weather station's ID number. The first two letters denote the country (using FIPS country codes)}
  \item{name}{The station's name}
  \item{latitude}{The station's latitude, in decimal degrees. Southern latitudes will be negative}
  \item{longitude}{The station's longitude, in decimal degrees. Western longitudes will be negative}
  \item{distance}{distance (km) from coordinates searched}
}
}
\source{
<https://www.ncdc.noaa.gov/cdo-web/webservices/v2>
}
\usage{
weather_stations_ghcnd
}
\description{
A table listing all NOAA GHCND weather stations within 100km of each prognosticator's city coordinates
}
\keyword{datasets}
