% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prognosticators}
\alias{prognosticators}
\title{Prognosticators data}
\format{
A data frame with 153 rows and 3 columns:
\describe{
  \item{prognosticator_name}{Prognosticator name}
  \item{prognosticator_type}{Prognosticator type, original labels}
  \item{prognosticator_city}{Where the prognosticator is located}
  \item{prognosticator_lat}{City latitude}
  \item{prognosticator_long}{City longitude}
  \item{prognosticator_status}{Prognosticator status: creature, human mascot, inanimate}
  \item{prognosticator_creature}{Prognosticator creature type}
  \item{prognosticator_phylum}{Prognosticator creature phylum}
  \item{prognosticator_class}{Prognosticator creature class}
  \item{prognosticator_order}{Prognosticator creature order}
}
}
\source{
<https://countdowntogroundhogday.com/groundhogs-from-around-the-world> "Countdown to Groundhog Day"
}
\usage{
prognosticators
}
\description{
A table listing all known February 2nd prognosticators
}
\keyword{datasets}
